a2a_mcp/requirements.txt
azure-functions
ddgs
fasta2a
nest_asyncio
pydantic-ai
pydantic-settings
pyyaml
requests
rich
uvicorn